+++ test/support/sqrt.c
@@ -1,5 +1,5 @@
 #include <math.h>
 
-double _mlir_ciface_Elixir$CtypesTest$SqrtTest$sqrt_c(double x) {
+double _mlir_ciface_Elixir$SqrtWrapper$sqrt_c(double x) {
     return sqrt(x);
 }
